+++ src/main/java/splitwise/service/divide/EqualStrategy.java
@@ -9,7 +9,7 @@ public class EqualStrategy implements DivideStrategy {
 	public final static String STRATEGY_NAME = "EQUAL";
 
 	@Override
-	public List<DividedAmount> divide(int amount, List<Integer> users, List<Double> division) {
+	public List<DividedAmount> divide(int amount, List<Integer> users, List<Integer> division) {
 
 		return null;
 	}
