+++ src/main/java/splitwise/service/divide/PercentStrategy.java
@@ -9,7 +9,7 @@ public class PercentStrategy implements DivideStrategy {
 	public final static String STRATEGY_NAME = "PERCENT";
 
 	@Override
-	public List<DividedAmount> divide(int amount, List<Integer> users, List<Double> division) {
+	public List<DividedAmount> divide(int amount, List<Integer> users, List<Integer> division) {
 
 		return null;
 	}
